+++ routes/home.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const router = express.Router();
+
+router.use(express.static('/public'));
+router.get('/', (req, res) => {
+  res.sendFile(__dirname + '/public/pages/index.html');
+});
+
+module.exports = router;
